Twitter/backend/requirements.txt
tweepy
requests_oauthlib
python-dotenv
Flask
